DEV-44: refractor, add comments, fix bug
