+++ routes/helperMethods.js
@@ -61,23 +61,28 @@ async function postNotification(message, user_id, quick_link_id, link_type) {
   return n;
 }
 
+// return an array of substrings of length 3+
 function ngram(query) {
   const MIN_LEN = 3; 
   const queryLen = query.length; 
   query = query.toLowerCase();
   const ngrams = new Set();
   ngrams.add(query);
+  // add substrings from largest to smallest 
   for (let substrLen = queryLen; substrLen >= MIN_LEN; substrLen--) {
     for (let i = 0; i <= queryLen - substrLen; i++) {
       ngrams.add(query.slice(i, i + substrLen)); 
     }
   }
+  // make array from set 
   return Array.from(ngrams);
 }
 
+// for strict query matching 
 async function simpleSearch(query, page) {
   const PERPAGE = 10; 
   const result = {}; 
+  // set pagination info; limit to 100 
   const total = await SISCV.countDocuments(query).exec(); 
   result.pagination = {
     page: page, 
@@ -85,11 +90,12 @@ async function simpleSearch(query, page) {
     last: total <= 100 ? Math.ceil(total / PERPAGE) : 10, 
     total: total
   }
-  let courses = await SISCV.find(query).skip((page) * PERPAGE).limit(PERPAGE); 
-  result.courses = courses; 
+  // skip and limit according to page 
+  result.courses = await SISCV.find(query).skip((page) * PERPAGE).limit(PERPAGE); 
   return result; 
 }
 
+// search for all courses matching substring of searchTerm  
 async function fuzzySearch(query, searchTerm, page) {
   const PERPAGE = 10; 
   const result = {};
@@ -98,11 +104,12 @@ async function fuzzySearch(query, searchTerm, page) {
     gram = gram.replace('.', '\\.'); 
     return new RegExp(gram);
   });
-  // TODO: index title and number 
+  // query for title / number matching any of the RegExps
   query['$or'] = [
     { title: { $in: regNgrams } },
     { number: { $in: regNgrams } },
   ]; 
+  // return pagination information; limit to 100 
   const total = await SISCV.countDocuments(query).exec(); 
   result.pagination = {
     page: page, 
@@ -110,7 +117,9 @@ async function fuzzySearch(query, searchTerm, page) {
     last: total <= 100 ? Math.ceil(total / PERPAGE) : 10, 
     total: total <= 100 ? total : 100, 
   }
+  // query for courses
   let courses = await SISCV.find(query); 
+  // calculate priority; summate the matching substring lengths 
   courses.forEach((course, i) => {
     for (let gram of ngrams) {
       if (course.title.toLowerCase().includes(gram) || course.number.toLowerCase().includes(gram)) {
@@ -118,20 +127,101 @@ async function fuzzySearch(query, searchTerm, page) {
       }
     }
   });
+  // sort by descending priority 
   courses = courses.sort((c1, c2) => c2.priority - c1.priority); 
+  // skip and limit according to page 
   result.courses = courses.slice(page * PERPAGE, (page + 1) * PERPAGE); 
   return result; 
 }
 
+// return true if userCourse is offered in newTerm 
 const checkDestValid = (sisCourses, userCourse, newTerm) => {
   for (let sisC of sisCourses) {
     if (sisC.number === userCourse.number) {
       for (let term of sisC.terms) {
+        // example: "Fall 2021" includes "Fall"
         if (term.includes(newTerm)) return true; 
       }
     }
   }
   return false; 
+}
+
+function constructQuery(params) {
+  let {
+    userQuery = "",
+    school = "",
+    department = "",
+    term = "",
+    areas = "",
+    wi,
+    credits = "",
+    tags = "",
+    level = "",
+  } = params;
+  userQuery = userQuery.replace(/[-\/\\^$*+?.()|[\]{}]/g, "\\$&"); //escape special character for regex
+  let query = {
+    $or: [
+      { title: { $regex: userQuery, $options: "i" } },
+      { number: { $regex: userQuery, $options: "i" } },
+    ],
+    "versions.school": { $regex: school, $options: "i" },
+    "versions.department": { $regex: department, $options: "i" },
+    "versions.term": { $regex: term, $options: "i" },
+    "versions.level": { $regex: level, $options: "i" },
+  };
+  if (areas !== "" && areas !== "None") {
+    query["versions.areas"] = {
+      $not: new RegExp("None"), 
+      $in: areas.split("|").map((area) => new RegExp(area)),
+    };
+  }
+
+  if (tags !== "") {
+    query["versions.tags"] = {
+      $in: tags.split("|").map((tag) => new RegExp(tag)),
+    };
+  }
+
+  if (credits !== "") {
+    query["versions.credits"] = {
+      $in: credits.split("|"),
+    };
+  }
+
+  if (wi != null) {
+    if (wi === "1" || wi === "true") {
+      query["versions.wi"] = true;
+    } else if (wi === "0" || wi === "false") {
+      query["versions.wi"] = false;
+    }
+  }
+  return query;
+}
+
+async function sendCourseVersion(query, version, res) {
+  const match = await SISCV.findOne(query); 
+  if (match == null) {
+    return errorHandler(
+      res,
+      404,
+      "Did not find any course or the course specified is not offered in this term."
+    );
+  } 
+  try {
+    let course = {};
+    course.title = match.title;
+    course.number = match.number;
+    course.terms = match.terms;
+    match.versions.forEach((v) => {
+      if (v.term === version) {
+        course.version = v;
+      }
+    });
+    returnData(course, res);
+  } catch (err) {
+    errorHandler(res, 400, err); 
+  }
 }
 
 export {
@@ -143,4 +233,6 @@ export {
   simpleSearch,
   fuzzySearch,
   checkDestValid,
+  constructQuery, 
+  sendCourseVersion
 };
